E-commerce-fin-beta

+++ src/components/ItemDetailContainer/ItemDetailContainer.jsx
@@ -1,19 +1,18 @@
-/* import { useEffect, useState } from "react"; */
 import { useState, useEffect } from "react";
-import { useParams } from "react-router-dom";
-import { getDoc, doc } from "firebase/firestore"; //Obtener un solo documento con getDoc // doc genera referencia al documento
+import { useParams, useNavigate } from "react-router-dom";
+import { getDoc, doc } from "firebase/firestore";
 import { db } from "../../../services/firebase/firebaseConfig";
 import PageTitle from "../PageTitle/PageTitle";
-/* import { getProductById } from "../AsyncMock/asyncMock"; */
 import ItemDetail from "../ItemDetail/ItemDetail";
-/* import { useAsync } from "../../Hooks/useAsync"; */
 import Loading from "../Loading/Loading";
-/* import RelatedProducts from "../RelatedProducts/RelatedProducts"; */
 
 const ItemDetailContainer = () => {
   const [product, setProduct] = useState(null);
   const [loading, setLoading] = useState();
+  const [productNotFound, setProductNotFound] = useState(false);
+
   const { itemId } = useParams();
+  const navigate = useNavigate();
 
   useEffect(() => {
     setLoading(true);
@@ -21,28 +20,30 @@ const ItemDetailContainer = () => {
 
     getDoc(productRef)
       .then((querySnapshot) => {
-        const fields = querySnapshot.data(); //almacenamos los datos de todos los campos de cada dato
-        const productAdapted = { id: querySnapshot.id, ...fields }; //retornamos un objetos con todos los datos, el id por separado ya que no se enceuntra dentro de data, esta separado
-        setProduct(productAdapted);
+        if (querySnapshot.size === 0) {
+          throw new Error("Error 404: not found");
+        }
+        if (querySnapshot.exists()) {
+          const fields = querySnapshot.data();
+          const productAdapted = { id: querySnapshot.id, ...fields };
+          setProduct(productAdapted);
 
-        localStorage.setItem("currentId", itemId);
+          localStorage.setItem("currentId", itemId);
+        } else {
+          setProductNotFound(true);
+        }
+      })
+      .catch((error) => {
+        return error;
       })
       .finally(() => {
         setLoading(false);
       });
   }, [itemId]);
 
-  /* loading === true && (
-    <div className="style pt-2">
-      <Loading />
-    </div>
-  ); */
-  /* const {
-    data: product,
-    error,
-    loading,
-  } = useAsync(() => getProductById(itemId), [itemId]); */
-
+  if (productNotFound) {
+    return navigate("/*");
+  }
   return (
     <div>
       {loading ? (
@@ -53,10 +54,7 @@ const ItemDetailContainer = () => {
       ) : (
         <ItemDetail {...product} />
       )}
-
-      {/* <RelatedProducts {...product} /> */}
     </div>
   );
 };
-
 export default ItemDetailContainer;
